[Map] Consolidate marker generators into one unified file, rather than in a different package.
